==> 09_Parasite_GSEA_Visualizations.Rmd <==
```{r}
rm(list = ls())
```


```{r}
library(openxlsx)
library(tidyverse)
library(ggpubr)
library(ggforce)
library(reshape2)
library(ggpubr)
library(org.Mm.eg.db)
library(AnnotationDbi)
library(enrichplot)
library(clusterProfiler)
library(msigdbr)
library(forcats)
library(ggrepel)
library(GSVA)
library(SingleCellExperiment)
library(scran)
library(ReactomePA)
library(broom)
library(multcomp)
library(parallel)
```

```{r}
fit <- readRDS("/vol/ExtraVol/ParasiteDEX/parasite_glmGamPoi_fit_pseudobulked.RDS")
cnt <- fit[["data"]]@assays@data@listData[["counts"]]
sce <- SingleCellExperiment(assays = list(counts = cnt))
sce <- computeSumFactors(sce)
sce <- logNormCounts(sce)
normalized_counts <- logcounts(sce)

#Read the genomic feature table
genomic_features <- read.csv("/vol/ExtraVol/Genomic_Features.tsv", row.names = NULL, sep="\t")
genomic_features_unique <- genomic_features %>%
  dplyr::select(c(Symbol, Name)) %>%
  dplyr::distinct()
protein_coding_genes <- unique(subset(genomic_features, Gene.Type == "protein-coding")$Symbol)

normalized_counts <- subset(normalized_counts, rownames(normalized_counts) %in% protein_coding_genes)

#This is to add full names of the genes to the results
gene_metadata <- dplyr::select(genomic_features, c(Name, Symbol)) %>% 
  dplyr::rename("name" = "Symbol",
                "open_name" = "Name") %>%
  distinct()

entrezid_mapper <- AnnotationDbi::select(org.Mm.eg.db, keys=keys(org.Mm.eg.db), columns = c("SYMBOL", "ENTREZID")) %>% dplyr::rename("Symbol" = "SYMBOL")

normalized_counts_temp <- as.data.frame(normalized_counts) %>%
  dplyr::mutate(Symbol = rownames(normalized_counts))
normalized_counts_withENTREZID <- merge(normalized_counts_temp, entrezid_mapper, by="Symbol")
rownames(normalized_counts_withENTREZID) <- normalized_counts_withENTREZID$ENTREZID
normalized_counts_withENTREZID <- dplyr::select(normalized_counts_withENTREZID, -c(ENTREZID, Symbol))
rm(normalized_counts_temp)
```

### Read the GSEA results ###

```{r}
# enrichment_reactome_infection <- readRDS("/vol/ExtraVol/ParasiteDEX/GSEA/enrichment_reactome_infection.RDS")
# enrichment_reactome_infection_list <- read.xlsx("/vol/ExtraVol/ParasiteDEX/GSEA/enrichment_reactome_infection_list.xlsx")
# 
# enrichment_gobp_infection <- readRDS("/vol/ExtraVol/ParasiteDEX/GSEA/enrichment_gobp_infection.RDS")
# enrichment_gobp_infection_list <- read.xlsx("/vol/ExtraVol/ParasiteDEX/GSEA/enrichment_gobp_infection_list.xlsx")
```

```{r}
reactome_gson <- gson_Reactome("mouse")
gobp_gson <- gson_GO(OrgDb = "org.Mm.eg.db", ont = "BP")
gomf_gson <- gson_GO(OrgDb = "org.Mm.eg.db", ont = "MF")
```

```{r}
gsid2gene <- reactome_gson@gsid2gene
gsid2name <- reactome_gson@gsid2name
reactome_temp <- merge(gsid2gene, gsid2name, by="gsid")

gsid2gene <- gobp_gson@gsid2gene
gsid2name <- gobp_gson@gsid2name
gobp <- merge(gsid2gene, gsid2name, by="gsid")

gsid2gene <- gomf_gson@gsid2gene
gsid2name <- gomf_gson@gsid2name
gomf <- merge(gsid2gene, gsid2name, by="gsid")
```

```{r}
genelist_df <- read.xlsx("/vol/ExtraVol/in_vitro_DEGs_IL-13_vs_ust_fitted values_FC1_p0.05.xlsx")

IL13custom_UP <- subset(genelist_df, log2FC > 0) %>%
  dplyr::select(X1) %>%
  dplyr::rename("gene" = X1) %>%
  dplyr::mutate("name" = "IL13custom_UP",
                "gsid" = "CUSTOM",
                gene = as.character(gene))

IL13custom_DOWN <- subset(genelist_df, log2FC < 0) %>%
  dplyr::select(X1) %>%
  dplyr::rename("gene" = X1) %>%
  dplyr::mutate("name" = "IL13custom_DOWN",
                "gsid" = "CUSTOM",
                gene = as.character(gene))

IL13custom <- bind_rows(IL13custom_UP, IL13custom_DOWN)

reactome <- bind_rows(reactome_temp, IL13custom)
```

```{r}
reactome_list <- unlist(split(reactome, 
                              f = reactome$name), 
                        recursive = F)

gobp_list <- unlist(split(gobp, 
                              f = gobp$name), 
                        recursive = F)

gomf_list <- unlist(split(gomf, 
                              f = gomf$name), 
                        recursive = F)
```

```{r}
gsva_object_reactome <- gsvaParam(exprData = as.matrix(normalized_counts_withENTREZID), geneSets = reactome_list, minSize = 10)
gsva_res_reactome <- gsva(gsva_object_reactome, verbose=TRUE)

gsva_object_gobp <- gsvaParam(exprData = as.matrix(normalized_counts_withENTREZID), geneSets = gobp_list, minSize = 10)
gsva_res_gobp <- gsva(gsva_object_gobp, verbose=TRUE)

gsva_object_gomf <- gsvaParam(exprData = as.matrix(normalized_counts_withENTREZID), geneSets = gomf_list, minSize = 10)
gsva_res_gomf <- gsva(gsva_object_gomf, verbose=TRUE)
```
### Functions to sort and annotate ###

```{r}
# Function to assign shorter group names based on keywords
short_group_name <- function(x) {
  case_when(
    str_detect(x, "uninfected") ~ "uninfected",
    str_detect(x, "early") ~ "early",
    str_detect(x, "peak") ~ "peak",
    str_detect(x, "repair") ~ "repair",
    TRUE ~ "other"
  )
}

# Function to create an ordering factor
ordering_factor <- function(x) {
  case_when(
    str_detect(x, "uninfected") ~ "1_uninfected",
    str_detect(x, "early") ~ "2_early",
    str_detect(x, "peak") ~ "3_peak",
    str_detect(x, "repair") ~ "4_repair",
    TRUE ~ "5_other"
  )
}

# Define group colors
group_colors <- c(
  "uninfected" = "#6baed6",  # Darker Pastel Blue
  "early" = "#fd8d3c",       # Darker Pastel Orange
  "peak" = "#fb6a4a",        # Darker Pastel Red
  "repair" = "#74c476"       # Darker Pastel Green
)

# Define the order of PlotGroup levels
plot_group_levels <- c("uninfected", "early", "peak", "repair")

# Function to sanitize filenames
sanitize_filename <- function(filename) {
  # Replace spaces with underscores and remove other special characters
  gsub("[^A-Za-z0-9_]", "", gsub(" ", "_", filename))
}
```

```{r}
reactome_geneset_sizes <- reactome %>%
  dplyr::group_by(name) %>%
  dplyr::summarise(GeneSetSize = n()) %>%
  dplyr::rename("GeneSet" = "name")

gobp_geneset_sizes <- gobp %>%
  dplyr::group_by(name) %>%
  dplyr::summarise(GeneSetSize = n()) %>%
  dplyr::rename("GeneSet" = "name")

gomf_geneset_sizes <- gomf %>%
  dplyr::group_by(name) %>%
  dplyr::summarise(GeneSetSize = n()) %>%
  dplyr::rename("GeneSet" = "name")
```


### Prepare GSVA result data frames for visualization ###

```{r}
### Prepare Reactome results ###

gsva_res_reactome_melted <- reshape2::melt(gsva_res_reactome, 
                                           varnames=c('GeneSet', 'SampleName'),
                                           value.name = "GSVAscore") %>%
  dplyr::mutate(SampleGroup = str_remove_all(SampleName, '_Rep.+'),
                GeneSet = str_remove_all(GeneSet, ".gene"),
                GeneSet = str_remove_all(GeneSet, "REACTOME_"),
                PlotGroup = short_group_name(SampleGroup),
                identity_simplified = str_extract(SampleGroup, "^[^_]+"))

# Add the ordering factor to the dataframe
gsva_res_reactome_melted <- gsva_res_reactome_melted %>%
  mutate(OrderFactor = ordering_factor(SampleGroup),
         SampleGroup = factor(SampleGroup, levels = unique(SampleGroup[order(OrderFactor)])))

# Optionally, you can remove the OrderFactor column if it's no longer needed
gsva_res_reactome_melted <- dplyr::select(gsva_res_reactome_melted, -OrderFactor)
saveRDS(gsva_res_reactome_melted, "/vol/ExtraVol/ParasiteDEX/GSEA/gsva_res_reactome_melted.RDS")

### Prepare GOBP results ###

gsva_res_gobp_melted <- reshape2::melt(gsva_res_gobp, 
                                           varnames=c('GeneSet', 'SampleName'),
                                           value.name = "GSVAscore") %>%
  dplyr::mutate(SampleGroup = str_remove_all(SampleName, '_Rep.+'),
                GeneSet = str_remove_all(GeneSet, ".gene"),
                PlotGroup = short_group_name(SampleGroup),
                identity_simplified = str_extract(SampleGroup, "^[^_]+"))

# Add the ordering factor to the dataframe
gsva_res_gobp_melted <- gsva_res_gobp_melted %>%
  mutate(OrderFactor = ordering_factor(SampleGroup),
         SampleGroup = factor(SampleGroup, levels = unique(SampleGroup[order(OrderFactor)])))

# Optionally, you can remove the OrderFactor column if it's no longer needed
gsva_res_gobp_melted <- dplyr::select(gsva_res_gobp_melted, -OrderFactor)
saveRDS(gsva_res_gobp_melted, "/vol/ExtraVol/ParasiteDEX/GSEA/gsva_res_gobp_melted.RDS")

### Prepare GOMF results ###

gsva_res_gomf_melted <- reshape2::melt(gsva_res_gomf, 
                                           varnames=c('GeneSet', 'SampleName'),
                                           value.name = "GSVAscore") %>%
  dplyr::mutate(SampleGroup = str_remove_all(SampleName, '_Rep.+'),
                GeneSet = str_remove_all(GeneSet, ".gene"),
                PlotGroup = short_group_name(SampleGroup),
                identity_simplified = str_extract(SampleGroup, "^[^_]+"))

# Add the ordering factor to the dataframe
gsva_res_gomf_melted <- gsva_res_gomf_melted %>%
  mutate(OrderFactor = ordering_factor(SampleGroup),
         SampleGroup = factor(SampleGroup, levels = unique(SampleGroup[order(OrderFactor)])))

# Optionally, you can remove the OrderFactor column if it's no longer needed
gsva_res_gomf_melted <- dplyr::select(gsva_res_gomf_melted, -OrderFactor)
saveRDS(gsva_res_gomf_melted, "/vol/ExtraVol/ParasiteDEX/GSEA/gsva_res_gomf_melted.RDS")

```

### ANOVA ###

```{r}
### Reactome ###

# Perform the two-way ANOVA analysis and summarize the results
anova_results_twoway_reactome <- gsva_res_reactome_melted %>%
  group_by(GeneSet) %>%
  summarize(
    aov_result = list(aov(GSVAscore ~ identity_simplified * PlotGroup, data = cur_data())),
    aov_tidy = list(tidy(aov(GSVAscore ~ identity_simplified * PlotGroup, data = cur_data()))),
    p_value_identity = aov_tidy[[1]] %>% filter(term == "identity_simplified") %>% pull(p.value),
    p_value_stage = aov_tidy[[1]] %>% filter(term == "PlotGroup") %>% pull(p.value),
    p_value_interaction = aov_tidy[[1]] %>% filter(term == "identity_simplified:PlotGroup") %>% pull(p.value)
  ) %>%
  dplyr::select(-aov_result, -aov_tidy) %>%
  dplyr::mutate(GeneSet = stringr::str_replace_all(GeneSet, ".gsid", "")) %>%
  dplyr::inner_join(reactome_geneset_sizes, by="GeneSet") %>%  
  dplyr::arrange(p_value_interaction)

# Perform the one-way ANOVA analysis and summarize the results
anova_results_oneway_reactome <- gsva_res_reactome_melted %>%
  group_by(GeneSet, identity_simplified) %>%
  summarize(
    aov_result = list(aov(GSVAscore ~ SampleGroup, data = cur_data())),
    aov_tidy = list(tidy(aov(GSVAscore ~ SampleGroup, data = cur_data()))),
    p_value_CellTypeVar = aov_tidy[[1]] %>% filter(term == "SampleGroup") %>% pull(p.value)
  ) %>%
  dplyr::select(-aov_result, -aov_tidy) %>%
  dplyr::mutate(GeneSet = stringr::str_replace_all(GeneSet, ".gsid", "")) %>%
  dplyr::inner_join(reactome_geneset_sizes, by="GeneSet")

anova_results_reactome <- merge(anova_results_twoway_reactome, anova_results_oneway_reactome, by="GeneSet")
anova_results_reactome <- anova_results_reactome %>% dplyr::arrange(p_value_identity, GeneSet, p_value_CellTypeVar)

# gsva_res_reactome_melted <- gsva_res_reactome_melted %>%
#   mutate(identity_simplified = as.factor(identity_simplified),
#          PlotGroup = as.factor(PlotGroup))

### GOBP ###

# Perform the two-way ANOVA analysis and summarize the results
anova_results_twoway_gobp <- gsva_res_gobp_melted %>%
  group_by(GeneSet) %>%
  summarize(
    aov_result = list(aov(GSVAscore ~ identity_simplified * PlotGroup, data = cur_data())),
    aov_tidy = list(tidy(aov(GSVAscore ~ identity_simplified * PlotGroup, data = cur_data()))),
    p_value_identity = aov_tidy[[1]] %>% filter(term == "identity_simplified") %>% pull(p.value),
    p_value_stage = aov_tidy[[1]] %>% filter(term == "PlotGroup") %>% pull(p.value),
    p_value_interaction = aov_tidy[[1]] %>% filter(term == "identity_simplified:PlotGroup") %>% pull(p.value)
  ) %>%
  dplyr::select(-aov_result, -aov_tidy) %>%
  dplyr::mutate(GeneSet = stringr::str_replace_all(GeneSet, ".gsid", "")) %>%
  dplyr::inner_join(gobp_geneset_sizes, by="GeneSet") %>%  
  dplyr::arrange(p_value_interaction)

# Perform the one-way ANOVA analysis and summarize the results
anova_results_oneway_gobp <- gsva_res_gobp_melted %>%
  group_by(GeneSet, identity_simplified) %>%
  summarize(
    aov_result = list(aov(GSVAscore ~ SampleGroup, data = cur_data())),
    aov_tidy = list(tidy(aov(GSVAscore ~ SampleGroup, data = cur_data()))),
    p_value_CellTypeVar = aov_tidy[[1]] %>% filter(term == "SampleGroup") %>% pull(p.value)
  ) %>%
  dplyr::select(-aov_result, -aov_tidy) %>%
  dplyr::mutate(GeneSet = stringr::str_replace_all(GeneSet, ".gsid", "")) %>%
  dplyr::inner_join(gobp_geneset_sizes, by="GeneSet")

anova_results_gobp <- merge(anova_results_twoway_gobp, anova_results_oneway_gobp, by="GeneSet")
anova_results_gobp <- anova_results_gobp %>% dplyr::arrange(p_value_identity, GeneSet, p_value_CellTypeVar)

### GOMF ###

# Perform the two-way ANOVA analysis and summarize the results
anova_results_twoway_gomf <- gsva_res_gomf_melted %>%
  group_by(GeneSet) %>%
  summarize(
    aov_result = list(aov(GSVAscore ~ identity_simplified * PlotGroup, data = cur_data())),
    aov_tidy = list(tidy(aov(GSVAscore ~ identity_simplified * PlotGroup, data = cur_data()))),
    p_value_identity = aov_tidy[[1]] %>% filter(term == "identity_simplified") %>% pull(p.value),
    p_value_stage = aov_tidy[[1]] %>% filter(term == "PlotGroup") %>% pull(p.value),
    p_value_interaction = aov_tidy[[1]] %>% filter(term == "identity_simplified:PlotGroup") %>% pull(p.value)
  ) %>%
  dplyr::select(-aov_result, -aov_tidy) %>%
  dplyr::mutate(GeneSet = stringr::str_replace_all(GeneSet, ".gsid", "")) %>%
  dplyr::inner_join(gomf_geneset_sizes, by="GeneSet") %>%
  dplyr::arrange(p_value_interaction)

# Perform the one-way ANOVA analysis and summarize the results
anova_results_oneway_gomf <- gsva_res_gomf_melted %>%
  group_by(GeneSet, identity_simplified) %>%
  summarize(
    aov_result = list(aov(GSVAscore ~ SampleGroup, data = cur_data())),
    aov_tidy = list(tidy(aov(GSVAscore ~ SampleGroup, data = cur_data()))),
    p_value_CellTypeVar = aov_tidy[[1]] %>% filter(term == "SampleGroup") %>% pull(p.value)
  ) %>%
  dplyr::select(-aov_result, -aov_tidy) %>%
  dplyr::mutate(GeneSet = stringr::str_replace_all(GeneSet, ".gsid", "")) %>%
  dplyr::inner_join(gomf_geneset_sizes, by="GeneSet")

anova_results_gomf <- merge(anova_results_twoway_gomf, anova_results_oneway_gomf, by="GeneSet")
anova_results_gomf <- anova_results_gomf %>% dplyr::arrange(p_value_identity, GeneSet, p_value_CellTypeVar)
```

```{r}
# Define the suffixes and file names
suffixes <- c("reactome", "gobp", "gomf")
file_names <- c("ANOVA_CellTypeSpecific_reactome.xlsx", "ANOVA_CellTypeSpecific_gobp.xlsx", "ANOVA_CellTypeSpecific_gomf.xlsx")

# Loop through each suffix and file name
for (j in seq_along(suffixes)) {
  suffix <- suffixes[j]
  file_name <- file_names[j]
  
  # Get the data frame based on the suffix
  df_name <- paste0("anova_results_oneway_", suffix)
  df <- get(df_name)
  
  # Get unique identities
  unique_identities <- unique(df$identity_simplified)
  
  # Create a new workbook
  wb <- createWorkbook()
  
  # Loop through each unique identity
  for (i in unique_identities) {
    # Split the data frame by identity_simplified
    df_split <- df[df$identity_simplified == i, ]
    
    # Sort the split data frame by p_value_CellTypeVar
    df_split_sorted <- df_split %>% arrange(p_value_CellTypeVar)
    
    # Add a new worksheet for the current identity
    addWorksheet(wb, sheetName = i)
    
    # Write the sorted data frame to the corresponding worksheet
    writeData(wb, sheet = i, x = df_split_sorted)
  }
  
  # Save the workbook to an Excel file
  saveWorkbook(wb, file.path("/vol/ExtraVol/ParasiteDEX/GSEA", file_name), overwrite = TRUE)
}
```

```{r}
# Create a new workbook
wb <- createWorkbook()

# Write anova_results_twoway_reactome to the "reactome" sheet
addWorksheet(wb, sheetName = "reactome")
writeData(wb, sheet = "reactome", x = anova_results_twoway_reactome)

# Write anova_results_twoway_gobp to the "gobp" sheet
addWorksheet(wb, sheetName = "gobp")
writeData(wb, sheet = "gobp", x = anova_results_twoway_gobp)

# Write anova_results_twoway_gomf to the "gomf" sheet
addWorksheet(wb, sheetName = "gomf")
writeData(wb, sheet = "gomf", x = anova_results_twoway_gomf)

# Save the workbook to an Excel file
saveWorkbook(wb, "/vol/ExtraVol/ParasiteDEX/GSEA/ANOVA_TwoWay.xlsx", overwrite = TRUE)
```

### GSVA PLOTS ###

```{r}
# Define unique gene sets
gene_sets <- unique(gsva_res_reactome_melted$GeneSet)


# Loop through each unique gene set
for (gene_set in gene_sets) {
  tryCatch({
    # Subset the data for the current gene set
    plotdata <- subset(gsva_res_reactome_melted, GeneSet == gene_set)
    plotdata$PlotGroup <- factor(plotdata$PlotGroup, levels = plot_group_levels)
    
    # Calculate the mean GSVAscore for each PlotGroup and identity_simplified
    mean_data <- plotdata %>%
      group_by(identity_simplified, PlotGroup) %>%
      summarize(mean_GSVAscore = mean(GSVAscore, na.rm = TRUE), .groups = 'drop')
    
    # Define the output file name, sanitized to remove problematic characters
    output_file <- paste0("/vol/ExtraVol/ParasiteDEX/GSEA/GSVA_", sanitize_filename(gene_set), ".pdf")
    
    # Save the plot as a PDF
    pdf(output_file, width = 2, height = 6)
    print(
      ggplot(plotdata, aes(y = GSVAscore, x = PlotGroup, color = PlotGroup)) +
        geom_point() +
        geom_line(data = mean_data, aes(y = mean_GSVAscore, group = 1), color = "black") +
        scale_color_manual(values = group_colors) +
        theme_minimal() +
        theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 5),
              axis.text.y = element_text(hjust = 1, size = 5),
              strip.text = element_text(size = 4),
              axis.title.x = element_blank(),
              axis.title.y = element_text(size=10),
              plot.title = element_text(size = 5)) +
        facet_wrap(~identity_simplified, ncol = 2, nrow = 5) +
        ggtitle(paste0(gene_set))+
        guides(color = "none")  # Remove the legend
    )
    dev.off()
  }, error = function(e) {
    message("Error generating plot for gene set: ", gene_set)
    message("Error message: ", e$message)
    dev.off()
  })
}

```

```{r}
plotdata <- subset(gsva_res_gobp_melted, GeneSet == "otic vesicle morphsis.gsid")
plotdata$PlotGroup <- factor(plotdata$PlotGroup, levels = c("uninfected",
                                                            "early",
                                                            "peak",
                                                            "repair"))
# Calculate the median GSVAscore for each PlotGroup and identity_simplified
mean_data <- plotdata %>%
  group_by(identity_simplified, PlotGroup) %>%
  summarize(mean_GSVAscore = mean(GSVAscore, na.rm = TRUE), .groups = 'drop')

pdf("/vol/ExtraVol/ParasiteDEX/GSEA/GSVA_mock.pdf", width = 2, height = 6)
ggplot(plotdata, aes(y = GSVAscore, x = PlotGroup, color = PlotGroup)) +
  geom_point() +
  geom_line(data = mean_data, aes(y = mean_GSVAscore, group = 1), color = "black") +
  scale_color_manual(values = group_colors) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 5),
        strip.text = element_text(size = 4),
        axis.title.x = element_blank()) +
  facet_wrap(~identity_simplified, ncol = 2, nrow=5)+
  guides(color = "none")  # Remove the legend
dev.off()
```

```{r}
ggplot(head(enrichment_reactome_infection_list,n=15), aes(x=GeneRatio, y=Description))+
  geom_dotplot()
```

```{r}
pdf("/vol/ExtraVol/ParasiteDEX/GSEA/dotplot_reactome.pdf", width=6, height = 7)
dotplot(enrichment_reactome_infection, showCategory=15, font.size=5)
dev.off()
```

```{r}
pdf("/vol/ExtraVol/ParasiteDEX/GSEA/dotplot_gobp.pdf", width=6, height = 7)
dotplot(enrichment_gobp_infection, showCategory=15, font.size=5)
dev.off()
```
