==> 10_Parasite_Visualizations.Rmd <==
```{r}
rm(list = ls())
```

```{r}
library(tidyverse)
library(data.table)
library(DESeq2)
library(conflicted)
library(openxlsx)
library(readxl)
library(tidyr)
#library(clusterProfiler)
library(ComplexHeatmap)
library(circlize)
library(forcats)
library(RColorBrewer)
library(reshape2)
options("print.matrix" = FALSE)
```

```{r}
coldata <- as.data.frame(readr::read_csv("/vol/ExtraVol/DataHistory/Parasite_Clustered_Metadata.csv")) %>% dplyr::select(CellName, identity, stage, stage_grouped)
rownames(coldata) <- coldata$CellName

#We have two leiden clusters of Adamdec1+ cells. We want to group them together for DEX analysis. We also have two samples of uninfected cells. We should group them too.
coldata <- coldata %>%
  dplyr::mutate(condition = paste0(identity, "_", stage_grouped)) %>%
  mutate(identity_simplified = gsub("\\+", "pos", identity)) %>%
  mutate(identity_simplified = gsub("\\-", "neg", identity_simplified)) %>%
  mutate(identity_simplified = gsub("\\s+", "", identity_simplified))

```

```{r}
cnt <- as.data.frame(readr::read_csv("/vol/ExtraVol/DataHistory/Parasite_Clustered_normalized.csv"))
rownames(cnt) <- cnt$CellName
cnt <- cnt %>% dplyr::select(-CellName)
cnt <- as.data.frame(t(cnt))
```

```{r}
representation <- dplyr::select(coldata, c(CellName, identity_simplified, stage_grouped)) %>%
  dplyr::group_by(identity_simplified, stage_grouped) %>%
  dplyr::summarise(n_cells = n()) %>%
  dplyr::ungroup() %>%
  dplyr::group_by(identity_simplified) %>%
  dplyr::mutate(pct_cells = (n_cells*100)/sum(n_cells),
                n_cells_total = sum(n_cells)) %>%
  dplyr::ungroup() %>%
  dplyr::arrange(desc(n_cells_total))

representation$stage_grouped <- factor(representation$stage_grouped, levels=c("uninfected_adult", "early", "peak", "repair"))
representation$identity_simplified <- factor(representation$identity_simplified, levels=unique(representation$identity_simplified))

# Combine the data for both plots
representation_long <- representation %>%
  gather(key = "measure", value = "value", pct_cells, n_cells_total)
representation_long$measure <- factor(representation_long$measure, levels = c("pct_cells","n_cells_total"))
representation_long$stage_grouped <- fct_rev(factor(representation_long$stage_grouped, levels=c("uninfected_adult", "early", "peak", "repair")))
```

### CALCULATING PERCENTAGE OF CELLS THAT EXPRESS THE MARKERS. RUN THIS ONLY ONCE, THEN READ THE OUTPUT. ###

```{r}
# # Read the Excel file
# file_path <- "/vol/ExtraVol/ParasiteDEX/ParasiteDEX_PairwiseMarkerResults.xlsx"
# sheet_names <- excel_sheets(file_path)
# 
# # Load each sheet as a separate data frame
# data_frames <- lapply(sheet_names, function(sheet) {
#   read_excel(file_path, sheet = sheet)
# })
# names(data_frames) <- sheet_names
# 
# # Initialize a list to store the results
# results <- list()
# 
# # Process each data frame
# for (sheet_name in sheet_names) {
#   # Get the corresponding data frame
#   df <- data_frames[[sheet_name]]
#   
#   # Subset coldata for the given identity_simplified
#   subset_coldata <- subset(coldata, identity_simplified == sheet_name & 
#                              stage_grouped == "uninfected_adult") 
#   subset_cells <- subset_coldata %>%
#     dplyr::pull(CellName)
#   
#   # Subset the cnt matrix to keep only the cells for the given identity
#   subset_cnt <- cnt[, subset_cells]
#   
#   # Calculate the percentage of cells with non-zero counts for each gene
#   gene_counts <- apply(subset_cnt, 1, function(row) {
#     sum(row > 0) / length(row) * 100
#   })
#   
#   # Store the result
#   results[[sheet_name]] <- data.frame(Symbol = rownames(subset_cnt), Percentage = gene_counts)
# }
# 
# # Print the results for verification
# print(results)
```

```{r}
# # Initialize a list to store the merged data frames
# merged_data_frames <- list()
# 
# # Merge corresponding data frames by the "Symbol" column
# for (sheet_name in sheet_names) {
#   df1 <- data_frames[[sheet_name]]
#   df2 <- results[[sheet_name]]
#   
#   # Perform the merge
#   merged_df <- df1 %>%
#     inner_join(df2, by = "Symbol")
#   
#   # Store the merged data frame
#   merged_data_frames[[sheet_name]] <- merged_df
# }
```

```{r}
# # Create a new workbook
# wb <- createWorkbook()
# 
# # Add each merged data frame as a separate sheet
# for (sheet_name in names(merged_data_frames)) {
#   addWorksheet(wb, sheet_name)
#   
#   # Write the data
#   writeData(wb, sheet = sheet_name, merged_data_frames[[sheet_name]])
# }
# 
# # Save the workbook
# saveWorkbook(wb, "/vol/ExtraVol/ParasiteDEX/ParasiteDEX_PairwiseMarkerResults_withPercentages.xlsx", overwrite = TRUE)
```

### END OF MARKER PERCENTAGE CALCULATION ###

```{r}
# Read the Excel file
file_path <- "/vol/ExtraVol/ParasiteDEX/ParasiteDEX_PairwiseMarkerResults_withPercentages.xlsx"
sheet_names <- excel_sheets(file_path)

# Load each sheet as a separate data frame
data_frames <- lapply(sheet_names, function(sheet) {
  read_excel(file_path, sheet = sheet)
})
names(data_frames) <- sheet_names
```

```{r}
# Initialize an empty vector to store the marker genes
markers <- c()

# Loop over each sheet name to process the corresponding data frame
for (sheet_name in sheet_names) {
  # Merge the corresponding data frames by the "Symbol" column
  df <- data_frames[[sheet_name]]
  
  # Filter genes by percentage expression and select the top marker genes
  top_markers <- df %>%
    subset(Percentage > 10) %>%  # Filter for genes with >10% non-zero expression
    dplyr::arrange(desc(signed_f_statistic)) %>%  # Sort by signed F-statistic
    head(10) %>%  # Select the top 10 marker genes
    dplyr::pull(Symbol) # Extract the "Symbol" column

  
  # Append the marker genes to the vector
  markers <- c(markers, top_markers)
}

markers <- unique(markers)
markers <- factor(markers, levels = markers)
sheet_names <- factor(sheet_names, levels = sheet_names)

# Print the markers vector for verification
print(markers)
```

```{r}
genes_list <- c("Pdpn", "Pdgfra", "Igfbp3", "Fgfr2", "Pdgfrb", "Des", "Acta2", "Myh11", "Actg2", "Cd81", "Ackr4", "Cd55", "Cd34", "Jam2", "Ccl11", "Kit", "Hand1")

uninfected_cells_df <- coldata %>%
  subset(stage_grouped == "uninfected_adult")

uninfected_cells_df$identity_simplified <- factor(uninfected_cells_df$identity_simplified, levels=unique(representation$identity_simplified))

# Sort the dataframe by the factor levels
uninfected_cells_df <- uninfected_cells_df %>%
  arrange(identity_simplified)

uninfected_cells <- uninfected_cells_df$CellName

cnt_markers <- subset(cnt, rownames(cnt) %in% genes_list)
cnt_markers <- cnt_markers %>% dplyr::select(all_of(uninfected_cells))

cnt_markers_temp <- reshape2::melt(as.matrix(cnt_markers), value.name = "Expression", varnames = c("Symbol", "CellName"))
cnt_markers_long <- merge(cnt_markers_temp, coldata, by="CellName") %>% 
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::mutate(MeanExpression = mean(Expression)) %>%
  dplyr::ungroup()

cnt_markers_percentages <- cnt_markers_long %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::summarise(
    nonzero_count_cells = sum(Expression != 0),
    total_count_cells = n()
  ) %>%
  dplyr::ungroup() %>%
  dplyr::mutate(pct_expressed = (nonzero_count_cells*100)/total_count_cells)

cnt_markers_scaled <- t(scale(t(cnt_markers), scale=FALSE))
cnt_markers_scaled_temp <- reshape2::melt(as.matrix(cnt_markers_scaled), value.name = "ScaledExpression", varnames = c("Symbol", "CellName"))
cnt_markers_nonscaled_temp <- reshape2::melt(as.matrix(cnt_markers), value.name = "Expression", varnames = c("Symbol", "CellName"))

cnt_markers_scaled_long <- merge(cnt_markers_scaled_temp, coldata, by="CellName") %>% 
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::summarize(MeanScaledExpression = mean(ScaledExpression),
                   Symbol = dplyr::first(Symbol),
                   identity_simplified = dplyr::first(identity_simplified)) %>%
  dplyr::ungroup()

cnt_markers_nonscaled_long <- merge(cnt_markers_temp, coldata, by="CellName") %>% 
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::summarize(MeanExpression = mean(Expression)) %>%
  dplyr::ungroup() 

cnt_markers_merged_temp <- merge(cnt_markers_scaled_long, cnt_markers_percentages, by="identity_Symbol")
cnt_markers_merged <- merge(cnt_markers_merged_temp, cnt_markers_nonscaled_long, by="identity_Symbol") %>%
  dplyr::distinct()
```

```{r}
ggplot(cnt_markers_merged, aes(x=Symbol, y=identity_simplified))+
  geom_count(aes(size=pct_expressed, color=MeanScaledExpression))+
  scale_color_gradient2(
    low = "blue", 
    mid = "white", 
    high = "red", 
    midpoint = 0, 
    name = "Scaled Mean Expression"
  ) +
  theme_minimal() +
  labs(
    x = "Gene",
    y = "Identity",
    size = "Percent Expressed"
  ) +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 45, hjust = 1)
  )

colors <- colorRamp2(c(0, 0.025 * max(cnt_markers_merged$MeanExpression), max(cnt_markers_merged$MeanExpression)),
                     c("grey100", "#add8e6", "#000080"))
cnt_markers_merged$identity_simplified <- factor(cnt_markers_merged$identity_simplified, levels = rev(sheet_names))

ggplot(cnt_markers_merged, aes(x=Symbol, y=identity_simplified))+
  geom_count(aes(size=pct_expressed, color=MeanExpression))+
  theme_minimal() +
  labs(
    x = "Gene",
    y = "Identity",
    size = "Percent Expressed"
  ) +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 45, hjust = 1),
    panel.grid.major = element_blank(),  # Remove major grid lines
    panel.grid.minor = element_blank()   # Remove minor grid lines
  ) +
  scale_color_gradientn(colors = colors(seq(0, max(cnt_markers_merged$MeanExpression), length.out = 100)))
```

```{r}
uninfected_cells_df <- coldata %>%
  subset(stage_grouped == "uninfected_adult")

uninfected_cells_df$identity_simplified <- factor(uninfected_cells_df$identity_simplified, levels=unique(representation$identity_simplified))

# Sort the dataframe by the factor levels
uninfected_cells_df <- uninfected_cells_df %>%
  arrange(identity_simplified)

uninfected_cells <- uninfected_cells_df$CellName

cnt_markers <- subset(cnt, rownames(cnt) %in% markers)
cnt_markers <- cnt_markers %>% dplyr::select(all_of(uninfected_cells))

cnt_markers_temp <- reshape2::melt(as.matrix(cnt_markers), value.name = "Expression", varnames = c("Symbol", "CellName"))
cnt_markers_long <- merge(cnt_markers_temp, coldata, by="CellName") %>% 
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::mutate(MeanExpression = mean(Expression)) %>%
  dplyr::ungroup()

cnt_markers_percentages <- cnt_markers_long %>%
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::summarise(
    nonzero_count_cells = sum(Expression != 0),
    total_count_cells = n()
  ) %>%
  dplyr::ungroup() %>%
  dplyr::mutate(pct_expressed = (nonzero_count_cells*100)/total_count_cells)

cnt_markers_scaled <- t(scale(t(cnt_markers), scale=FALSE))
cnt_markers_scaled_temp <- reshape2::melt(as.matrix(cnt_markers_scaled), value.name = "ScaledExpression", varnames = c("Symbol", "CellName"))
cnt_markers_nonscaled_temp <- reshape2::melt(as.matrix(cnt_markers), value.name = "Expression", varnames = c("Symbol", "CellName"))


cnt_markers_scaled_long <- merge(cnt_markers_scaled_temp, coldata, by="CellName") %>% 
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::summarize(MeanScaledExpression = mean(ScaledExpression),
                   Symbol = dplyr::first(Symbol),
                   identity_simplified = dplyr::first(identity_simplified)) %>%
  dplyr::ungroup()

cnt_markers_nonscaled_long <- merge(cnt_markers_temp, coldata, by="CellName") %>% 
  dplyr::mutate(identity_Symbol = paste0(identity_simplified, "_", Symbol)) %>%
  dplyr::group_by(identity_Symbol) %>%
  dplyr::summarize(MeanExpression = mean(Expression)) %>%
  dplyr::ungroup() 

cnt_markers_merged_temp <- merge(cnt_markers_scaled_long, cnt_markers_percentages, by="identity_Symbol")
cnt_markers_merged <- merge(cnt_markers_merged_temp, cnt_markers_nonscaled_long, by="identity_Symbol") %>%
  dplyr::distinct()
```

```{r}
ggplot(cnt_markers_merged, aes(x=Symbol, y=identity_simplified))+
  geom_count(aes(size=pct_expressed, color=MeanScaledExpression))+
  scale_color_gradient2(
    low = "blue", 
    mid = "white", 
    high = "red", 
    midpoint = 0, 
    name = "Scaled Mean Expression"
  ) +
  theme_minimal() +
  labs(
    x = "Gene",
    y = "Identity",
    size = "Percent Expressed"
  ) +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 45, hjust = 1)
  )

colors <- colorRamp2(c(0, 0.025 * max(cnt_markers_merged$MeanExpression), max(cnt_markers_merged$MeanExpression)),
                     c("grey100", "#add8e6", "#000080"))
cnt_markers_merged$Symbol <- factor(cnt_markers_merged$Symbol, levels = markers)
cnt_markers_merged$identity_simplified <- factor(cnt_markers_merged$identity_simplified, levels = rev(sheet_names))

ggplot(cnt_markers_merged, aes(x=Symbol, y=identity_simplified))+
  geom_count(aes(size=pct_expressed, color=MeanExpression))+
  theme_minimal() +
  labs(
    x = "Gene",
    y = "Identity",
    size = "Percent Expressed"
  ) +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 45, hjust = 1),
    panel.grid.major = element_blank(),  # Remove major grid lines
    panel.grid.minor = element_blank()   # Remove minor grid lines
  ) +
  scale_color_gradientn(colors = colors(seq(0, max(cnt_markers_merged$MeanExpression), length.out = 100)))
```


```{r}
marker_genes_df <- read.xlsx("/vol/ExtraVol/ParasiteDEX/ParasiteDEX_Markers.xlsx")
marker_genes_temp <- marker_genes_df %>%
  dplyr::arrange(desc(f_statistic)) %>%
  head(n=100) %>%
  dplyr::pull(Symbol)
marker_genes <- c(marker_genes_temp, "Cd81", "Hand1", "Ccl11", "Cd55")

infection_genes_df <- read.xlsx("/vol/ExtraVol/ParasiteDEX/ParasiteDEX_MainInfectionResults.xlsx")
infection_genes <- infection_genes_df %>%
  subset(!str_detect(Symbol, "Rpl|Rps")) %>%
  dplyr::arrange(desc(f_statistic)) %>%
  head(n=100) %>%
  dplyr::pull(Symbol)

interaction_genes_df <- read.xlsx("/vol/ExtraVol/ParasiteDEX/ParasiteDEX_InteractionResults.xlsx")
interaction_genes <- interaction_genes_df %>%
  subset(!str_detect(Symbol, "Rpl|Rps")) %>%
  dplyr::arrange(desc(f_statistic)) %>%
  head(n=100) %>%
  dplyr::pull(Symbol)

uninfected_cells_df <- coldata %>%
  subset(stage_grouped == "uninfected_adult")

uninfected_cells_df$identity <- factor(uninfected_cells_df$identity, levels=unique(representation$identity))

# Sort the dataframe by the factor levels
uninfected_cells_df <- uninfected_cells_df %>%
  arrange(identity)

uninfected_cells <- uninfected_cells_df$CellName

cnt_markers <- subset(cnt, rownames(cnt) %in% marker_genes)
cnt_markers <- cnt_markers %>% dplyr::select(all_of(uninfected_cells))

cnt_infection <- subset(cnt, rownames(cnt) %in% infection_genes)

cnt_infection_temp <- reshape2::melt(as.matrix(cnt_infection), value.name = "Expression", varnames = c("Symbol", "CellName"))
cnt_infection_long <- merge(cnt_infection_temp, coldata, by="CellName") %>% 
  dplyr::mutate(condition_Symbol = paste0(condition, "_", Symbol)) %>%
  dplyr::group_by(condition_Symbol) %>%
  dplyr::mutate(MeanExpression = mean(Expression)) %>%
  dplyr::ungroup()

cnt_infection_percentages <- cnt_infection_long %>%
  dplyr::mutate(condition_Symbol = paste0(condition, "_", Symbol)) %>%
  dplyr::group_by(condition_Symbol) %>%
  dplyr::summarise(
    nonzero_count_cells = sum(Expression != 0),
    total_count_cells = n()
  ) %>%
  dplyr::ungroup() %>%
  dplyr::mutate(pct_expressed = (nonzero_count_cells*100)/total_count_cells)

cnt_infection_scaled <- t(scale(t(cnt_infection)))
cnt_infection_scaled_temp <- reshape2::melt(as.matrix(cnt_infection_scaled), value.name = "ScaledExpression", varnames = c("Symbol", "CellName"))
cnt_infection_scaled_long <- merge(cnt_infection_scaled_temp, coldata, by="CellName") %>% 
  dplyr::mutate(condition_Symbol = paste0(condition, "_", Symbol)) %>%
  dplyr::group_by(condition_Symbol) %>%
  dplyr::mutate(MeanScaledExpression = mean(ScaledExpression)) %>%
  dplyr::ungroup()

cnt_infection_merged <- merge(cnt_infection_scaled_long, cnt_infection_percentages, by="condition_Symbol") %>%
  dplyr::select(MeanScaledExpression, identity, condition, Symbol, condition_Symbol, pct_expressed) %>%
  dplyr::distinct()

cnt_interaction <- subset(cnt, rownames(cnt) %in% interaction_genes)

cnt_interaction_temp <- reshape2::melt(as.matrix(cnt_interaction), value.name = "Expression", varnames = c("Symbol", "CellName"))
cnt_interaction <- merge(cnt_interaction_temp, coldata, by="CellName")
cnt_interaction <- cnt_interaction %>% 
  dplyr::mutate(condition_Symbol = paste0(condition, "_", Symbol)) %>%
  dplyr::group_by(condition_Symbol) %>%
  dplyr::mutate(MeanExpression = mean(Expression)) %>%
  dplyr::ungroup()

cnt_interaction_percentages <- cnt_interaction %>%
  dplyr::mutate(condition_Symbol = paste0(condition, "_", Symbol)) %>%
  dplyr::group_by(condition_Symbol) %>%
  dplyr::summarise(
    nonzero_count_cells = sum(Expression != 0),
    total_count_cells = n()
  ) %>%
  dplyr::ungroup() %>%
  dplyr::mutate(pct_expressed = (nonzero_count_cells*100)/total_count_cells)

cnt_interaction_merged <- merge(cnt_interaction, cnt_interaction_percentages, by="condition_Symbol") %>%
  dplyr::select(MeanExpression, identity, condition, Symbol, condition_Symbol, pct_expressed) %>%
  dplyr::distinct()

```

```{r}
# Define the bolder pastel colors
stage_colors <- c("uninfected_adult" = "#A9A9A9",
                  "early" = "#FFA07A",
                  "peak" = "#FF6347",
                  "repair" = "#77DD77")
```

```{r}
ggplot(cnt_infection_merged, aes(x=condition, y=Symbol))+
  geom_count(aes(size=pct_expressed, color=MeanScaledExpression))+
  scale_color_gradient2(
    low = "blue", 
    mid = "white", 
    high = "red", 
    midpoint = 0, 
    name = "Mean Scaled Expression"
  ) +
  theme_minimal() +
  labs(
    x = "Condition",
    y = "Symbol",
    size = "Percent Expressed"
  ) +
  theme(
    legend.position = "right",
    axis.text.x = element_text(angle = 45, hjust = 1)
  )
```

```{r}
# Combined faceted plot

facet_labels <- c("pct_cells" = "% of Cells", "n_cells_total" = "Number of Cells")
legend_order <- levels(fct_rev(factor(representation_long$stage_grouped)))

pdf("/vol/ExtraVol/ParasiteDEX/representations_parasite.pdf", width=10, height=5)
ggplot(data = representation_long, aes(x = value, y = fct_rev(identity), fill = stage_grouped)) +
  geom_bar(stat = "identity") +
  scale_fill_manual(values = (stage_colors), breaks = legend_order) +
  facet_wrap(~ measure, scales = "free_x", labeller = as_labeller(facet_labels)) +
  theme_minimal(base_size = 15) +
  theme(
    panel.background = element_blank(),  # Remove panel background
    plot.background = element_blank(),  # Remove plot background
    panel.grid.major = element_blank(),  # Remove major grid lines
    panel.grid.minor = element_blank(),  # Remove minor grid lines
    axis.line = element_line(color = "black"),  # Add axis lines
    legend.position = "bottom",  # Place legend at the bottom
    strip.text = element_text(size = 15),  # Adjust facet label size
    panel.spacing = unit(0.2, "lines"),  # Adjust space between facets
    panel.spacing.x = unit(0.2, "cm"),  # Adjust horizontal space between facets
    strip.placement = "outside",  # Place facet labels outside the plot
    plot.margin = margin(10, 10, 10, 10),  # Adjust plot margins
    plot.tag = element_text(size = 15),  # Adjust size of plot tag
    panel.grid = element_blank(),  # Remove panel grid lines
    strip.background = element_blank(),  # Remove strip background
    plot.tag.position = "bottomright",     # Position of plot tag
    axis.title.x = element_blank(),
    axis.title.y = element_blank()
  ) +
  labs(fill = "Infection Stage")  # Label for the legend
dev.off()
```

```{r}
ordered_data <- as.matrix(cnt_markers[,uninfected_cells_df$CellName])

group_labels <- factor(uninfected_cells_df$identity)
group_labels_unique <- unique(group_labels)
 
group_colors <- c(
  "PdgfraloCd55+ FB" = "#a6cee3",   
  "Cd81+ FB" = "#b2df8a",           
  "Pdgfralo FB" = "#fdbf6f",        
  "Pdgfrahi FB" = "#ff7f00",        
  "Ccl11+ MF" = "#6a3d9a",          
  "PdgfrahiCd55+ FB" = "#fb9a99",   
  "PdgfrahiJam2+ FB" = "#1f78b4",   
  "Pericytes" = "#b15928",          
  "Ccl11- MF" = "#CAB2D6",    
  "Hand1+" = "#8c8c8c"              
)
names(group_colors) <- group_labels_unique

column_annotation <- HeatmapAnnotation(
  group = group_labels,
  col = list(group = group_colors)
)

heatmap_colors <- colorRamp2(c(0, max(cnt_markers)),
                             c("grey100", "#800000"))

pdf("/vol/ExtraVol/ParasiteDEX/Heatmap_ParasiteMarkers.pdf", width=12, height=16)
Heatmap(ordered_data,
        name = "Normalized Counts", 
        show_column_names = FALSE,
        col = heatmap_colors,
        top_annotation = column_annotation,
        cluster_columns = FALSE,
        cluster_rows = TRUE,
        use_raster = FALSE
        )
dev.off()
```

```{r}
# Generate a palette with 10 colors
palette <- brewer.pal(n = 10, name = "Paired")

# Print the colors to see them
print(palette)
barplot(rep(1, 10), col = palette, border = NA)
```
